make changes to overflow problem, int-long

+++ Master/src/main/scala/Master.scala
@@ -7,7 +7,7 @@ import java.security.MessageDigest
 
 trait Message
 case object Mining extends Message
-case class Work(uf: String, k: Int, start: Int, nrOfElements: Int) extends Message
+case class Work(uf: String, k: Int, start: Long, nrOfElements: Long) extends Message
 case class Result(bitcoin: String, hash: String) extends Message
 case object AskForTask extends Message
 case object Start extends Message
@@ -16,7 +16,7 @@ case object Stop extends Message
   
   class Worker extends Actor {
     
-	def map(num: Int):String = {
+	def map(num: Long):String = {
       var s = (32 + num % 95).toChar.toString
       if (num / 95 == 0)  s else s + map(num / 95)
     }
@@ -53,7 +53,7 @@ case object Stop extends Message
     
   }
   
-  class Master(uf: String, k: Int, nrOfWorkers: Int, nrOfMessages: Int, nrOfElements: Int) extends Actor {
+  class Master(uf: String, k: Int, nrOfWorkers: Int, nrOfMessages: Int, nrOfElements: Long) extends Actor {
     //val start: Long = System.currentTimeMillis
     var MessageSent = 0
     var stopnum = 0
